IOS-4171 User RoundedButtonView instead of Button itself
